assignment-grader/pyproject.toml
[project]
name = "assignment-grader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "fuzzywuzzy>=0.18.0",
    "google-genai>=1.12.1",
    "openai>=1.0.0",
    "pymupdf>=1.21.0",
    "pypdf2>=3.0.1",
    "python-docx>=0.8.11",
    "python-dotenv>=0.21.0",
    "python-levenshtein>=0.12.2",
    "requests>=2.28.0",
    "streamlit>=1.22.0",
    "uvicorn>=0.34.2",
]
